+++ test/Unit/PowerShell/TestScripts/testBasicFunctionWithTriggerMetadata.ps1
@@ -8,4 +8,4 @@ param ($Req, $TriggerMetadata)
 # Used for logging tests
 Write-Verbose "a log"
 
-Push-OutputBinding -Name res -Value $Req
+Push-OutputBinding -Name res -Value $TriggerMetadata
